Added classes in different files - created new processor file
